requirements.txt
ablog
docutils
myst_nb
MyST-Parser
pydata-sphinx-theme
sphinx
sphinx-design
sphinxcontrib-bibtex
sphinxext-opengraph
pyyaml
sphinx-autobuild
myst-parser